requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
Django==5.2.11
django-bootstrap5==25.3
django-cleanup==9.0.0
django-htmx==1.27.0
django-sass-processor==1.4.2
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
frozenlist==1.8.0
idna==3.11
libsass==0.23.0
Markdown==3.10
multidict==6.7.0
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pillow==12.0.0
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytils==0.4.4
pytz==2025.2
six==1.17.0
slugify==0.0.1
sqlparse==0.5.4
telepot==12.7
typus==0.2.2
tzdata==2025.3
urllib3==2.6.2
yarl==1.22.0
